refactor: :recycle: rename db.js to mongo.js
